Reorganize ProjectManager code to be more idiomatic
